Vectorize frame rasterization, ship uint8 frames, batch select_action
Issue #24 — worker rasterization:
- _draw_circle/_draw_thick_line per-pixel Python loops replaced by
  _disc_pixels (every disc of a pass rasterized with numpy) and one thick
  polyline per snake body. Output is bit-identical to the old loops (same
  sample points, same in-disc test, same head-then-body draw order).
- Food-string binning + 8-connected flood fill in food_sense is now numpy
  cell grouping + min-label connected components (same strings, order).
- One (n, 3) foods array per tick shared by matrix, sectors and step().
- Enemy body sector scoring, nearby_food_mass and _min_enemy_distance
  vectorized.
- Frames leave the env as uint8: 77 KB per frame over the worker pipe
  instead of 307 KB; VecFrameStack stacks stay uint8; the replay buffer
  stores them without re-quantizing.

Issue #23 — action selection:
- DDQNAgent.select_actions(states, ids): reflexes/epsilon per agent on
  the CPU, then one batched uint8 forward (pinned staging, one sync).
- The trainer decides all live agents in one call; cudnn.benchmark on CUDA.

Measured on an idle RTX 3090:
  _process_data_to_matrix @160px: 3.0->0.97 ms (200 food, 1 enemy),
  11.1->1.8 ms (800 food, 3x80), 15.9->2.6 ms (800 food, 6x150)
  locked_food_target @800 food: 1.46->0.39 ms
  action selection per tick: 2 agents 1.56->0.67 ms, 8 agents 6.4->1.0 ms

+++ worker_process.py
@@ -200,7 +200,7 @@ def echo_worker(remote, parent_remote, worker_id, headless, nickname, matrix_siz
             remote.send('ok')
         elif cmd == 'reset':
             remote.send({
-                'matrix': np.zeros((3, matrix_size, matrix_size), dtype=np.float32),
+                'matrix': np.zeros((3, matrix_size, matrix_size), dtype=np.uint8),
                 'sectors': np.zeros(99, dtype=np.float32),
             })
         elif cmd == 'reset_one':
@@ -233,7 +233,7 @@ def delayed_echo_worker(remote, parent_remote, worker_id, headless, nickname, ma
             remote.send('ok')
         elif cmd == 'reset':
             remote.send({
-                'matrix': np.zeros((3, matrix_size, matrix_size), dtype=np.float32),
+                'matrix': np.zeros((3, matrix_size, matrix_size), dtype=np.uint8),
                 'sectors': np.zeros(99, dtype=np.float32),
             })
         elif cmd == 'reset_one':
